backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.10.3
anthropic==0.39.0
lxml==5.3.0
python-multipart==0.0.18
python-dotenv==1.0.1
